+++ aiofence/contrib/__init__.py
@@ -2,14 +2,16 @@
 Framework integrations. Requires ``starlette``; never imported by the core package.
 """
 
-from .middleware import (
+from .starlette import (
     DISCONNECT_EVENT_SCOPE_KEY,
     DisconnectMiddleware,
     get_disconnect_event,
+    require_disconnect_event,
 )
 
 __all__ = [
     "DISCONNECT_EVENT_SCOPE_KEY",
     "DisconnectMiddleware",
     "get_disconnect_event",
+    "require_disconnect_event",
 ]
