+++ tests/contrib/test_fastapi.py
@@ -1,3 +1,10 @@
+"""
+The FastAPI aliases and the ``dependencies=[...]`` wiring around them.
+
+Every app here is a ``fenced_app`` — the dependencies read the event
+``DisconnectMiddleware`` publishes and raise without it.
+"""
+
 from __future__ import annotations
 
 import asyncio
@@ -5,28 +12,59 @@ import json
 from collections.abc import AsyncIterator
 from typing import Any
 
+import pytest
 from fastapi import APIRouter, Depends, FastAPI
 from starlette.requests import Request
 from starlette.responses import StreamingResponse
 
 from aiofence import Fencing, get_current_fencing
-from aiofence.contrib.fastapi import DisconnectEvent, DisconnectFencing
-from aiofence.contrib.starlette import disconnect_fencing, disconnect_fencing_dependency
+from aiofence.contrib.fastapi import (
+    DisconnectEvent,
+    DisconnectFencing,
+    disconnect_fencing,
+    disconnect_fencing_dependency,
+)
 
 from .asgi_harness import (
     bound_codes,
     call_app,
     call_app_body,
+    fenced_app,
     http_scope,
     post_scope,
     scripted_receive,
 )
 
+# --- DisconnectMiddleware is required ---
+
+
+async def test__disconnect_fencing__when_middleware_missing__then_request_raises() -> None:
+    app = FastAPI()
+
+    @app.get("/work", dependencies=[Depends(disconnect_fencing)])
+    async def work() -> dict[str, Any]:
+        return {}
+
+    with pytest.raises(RuntimeError, match="DisconnectMiddleware"):
+        await call_app(app)
+
+
+async def test__disconnect_event_alias__when_middleware_missing__then_request_raises() -> None:
+    app = FastAPI()
+
+    @app.get("/work")
+    async def work(gone: DisconnectEvent) -> dict[str, bool]:
+        return {"gone": gone.is_set()}
+
+    with pytest.raises(RuntimeError, match="DisconnectMiddleware"):
+        await call_app(app)
+
+
 # --- binding via FastAPI's dependencies=[...] ---
 
 
 async def test__route_dependencies__when_declared__then_fencing_bound_in_handler() -> None:
-    app = FastAPI()
+    app = fenced_app()
 
     @app.get("/work", dependencies=[Depends(disconnect_fencing)])
     async def work() -> dict[str, Any]:
@@ -36,7 +74,7 @@ async def test__route_dependencies__when_declared__then_fencing_bound_in_handler
 
 
 async def test__app_dependencies__when_declared__then_fencing_bound_in_handler() -> None:
-    app = FastAPI(dependencies=[Depends(disconnect_fencing)])
+    app = fenced_app(dependencies=[Depends(disconnect_fencing)])
 
     @app.get("/work")
     async def work() -> dict[str, Any]:
@@ -52,14 +90,14 @@ async def test__router_dependencies__when_custom_code__then_uses_it() -> None:
     async def work() -> dict[str, Any]:
         return {"codes": bound_codes()}
 
-    app = FastAPI()
+    app = fenced_app()
     app.include_router(router)
 
     assert await call_app(app) == {"codes": ["client_gone"]}
 
 
 async def test__route_dependencies__when_nested_callee__then_fencing_propagates() -> None:
-    app = FastAPI()
+    app = fenced_app()
 
     async def service() -> list[str | None]:
         return bound_codes()
@@ -73,7 +111,7 @@ async def test__route_dependencies__when_nested_callee__then_fencing_propagates(
 
 async def test__route_dependencies__when_declared_twice__then_solved_once() -> None:
     """One callable, one FastAPI cache key — this proves caching, not sharing."""
-    app = FastAPI()
+    app = fenced_app()
     solved: list[str] = []
 
     async def counted(request: Request) -> AsyncIterator[Fencing]:
@@ -94,7 +132,7 @@ async def test__route_dependencies__when_declared_twice__then_solved_once() -> N
 
 async def test__distinct_dependencies__when_declared__then_share_one_event() -> None:
     """Two callables FastAPI cannot collapse by cache key — real sharing, not caching."""
-    app = FastAPI()
+    app = fenced_app()
     client_gone = disconnect_fencing_dependency(code="client_gone")
 
     @app.get("/work", dependencies=[Depends(disconnect_fencing), Depends(client_gone)])
@@ -106,7 +144,7 @@ async def test__distinct_dependencies__when_declared__then_share_one_event() -> 
 
 
 async def test__distinct_dependencies__when_codes_differ__then_both_bound() -> None:
-    app = FastAPI()
+    app = fenced_app()
     client_gone = disconnect_fencing_dependency(code="client_gone")
 
     @app.get("/work", dependencies=[Depends(disconnect_fencing), Depends(client_gone)])
@@ -124,7 +162,7 @@ async def test__app_and_router_dependencies__when_codes_differ__then_both_bound(
     async def work() -> dict[str, Any]:
         return {"codes": bound_codes()}
 
-    app = FastAPI(dependencies=[Depends(disconnect_fencing)])
+    app = fenced_app(dependencies=[Depends(disconnect_fencing)])
     app.include_router(router)
 
     assert await call_app(app) == {"codes": ["client_gone", "disconnect"]}
@@ -143,7 +181,7 @@ async def test__app_and_router_dependencies__when_disconnect__then_both_codes_ca
             "client_gone": fence.cancelled_by("client_gone"),
         }
 
-    app = FastAPI(dependencies=[Depends(disconnect_fencing)])
+    app = fenced_app(dependencies=[Depends(disconnect_fencing)])
     app.include_router(router)
 
     result = await call_app(app, receive=scripted_receive({"type": "http.disconnect"}))
@@ -152,7 +190,7 @@ async def test__app_and_router_dependencies__when_disconnect__then_both_codes_ca
 
 
 async def test__aliases__when_both_declared__then_single_receive_loop() -> None:
-    app = FastAPI()
+    app = fenced_app()
     receive = scripted_receive()
 
     @app.get("/work")
@@ -164,7 +202,7 @@ async def test__aliases__when_both_declared__then_single_receive_loop() -> None:
 
 
 async def test__app_dependencies__when_mixed_with_aliases__then_one_shared_trigger() -> None:
-    app = FastAPI(dependencies=[Depends(disconnect_fencing)])
+    app = fenced_app(dependencies=[Depends(disconnect_fencing)])
 
     @app.get("/work")
     async def work(fencing: DisconnectFencing, gone: DisconnectEvent) -> dict[str, Any]:
@@ -177,7 +215,7 @@ async def test__app_dependencies__when_mixed_with_aliases__then_one_shared_trigg
 
 
 async def test__disconnect_fencing_alias__when_used__then_binds_fencing() -> None:
-    app = FastAPI()
+    app = fenced_app()
 
     @app.get("/work")
     async def work(fencing: DisconnectFencing) -> dict[str, Any]:
@@ -187,7 +225,7 @@ async def test__disconnect_fencing_alias__when_used__then_binds_fencing() -> Non
 
 
 async def test__disconnect_event_alias__when_client_disconnects__then_event_set() -> None:
-    app = FastAPI()
+    app = fenced_app()
 
     @app.get("/work")
     async def work(gone: DisconnectEvent) -> dict[str, bool]:
@@ -200,7 +238,7 @@ async def test__disconnect_event_alias__when_client_disconnects__then_event_set(
 
 
 async def test__disconnect_event_alias__when_client_stays__then_event_unset() -> None:
-    app = FastAPI()
+    app = fenced_app()
 
     @app.get("/work")
     async def work(gone: DisconnectEvent) -> dict[str, bool]:
@@ -214,7 +252,7 @@ async def test__disconnect_event_alias__when_client_stays__then_event_unset() ->
 
 
 async def test__route_dependencies__when_client_disconnects__then_fence_cancelled() -> None:
-    app = FastAPI()
+    app = fenced_app()
     observed: list[bool] = []
 
     @app.get("/work", dependencies=[Depends(disconnect_fencing)])
@@ -232,7 +270,7 @@ async def test__route_dependencies__when_client_disconnects__then_fence_cancelle
 
 
 async def test__route_dependencies__when_client_stays__then_fence_not_cancelled() -> None:
-    app = FastAPI()
+    app = fenced_app()
     observed: list[bool] = []
 
     @app.get("/work", dependencies=[Depends(disconnect_fencing)])
@@ -248,12 +286,12 @@ async def test__route_dependencies__when_client_stays__then_fence_not_cancelled(
     assert observed == [False]
 
 
-# --- raw body caveat ---
+# --- raw body ---
 
 
 async def test__route_dependencies__when_body_param_declared__then_raw_read_uses_cache() -> None:
-    """FastAPI caches the body before dependencies run, so the watcher never sees it."""
-    app = FastAPI()
+    """The middleware replays body messages, so a fenced route can still read them."""
+    app = fenced_app()
 
     @app.post("/work", dependencies=[Depends(disconnect_fencing)])
     async def work(payload: dict[str, int], request: Request) -> dict[str, Any]:
@@ -275,7 +313,7 @@ async def test__route_dependencies__when_body_param_declared__then_raw_read_uses
 
 async def test__route_dependencies__when_streaming_response__then_fencing_bound_in_body() -> None:
     """Yield dependencies tear down after the body streams, so the context outlives it."""
-    app = FastAPI()
+    app = fenced_app()
 
     @app.get("/work", dependencies=[Depends(disconnect_fencing)])
     async def work() -> StreamingResponse:
@@ -289,7 +327,7 @@ async def test__route_dependencies__when_streaming_response__then_fencing_bound_
 
 async def test__streaming_response__when_server_owns_no_reader__then_fence_cancelled() -> None:
     """ASGI spec_version 2.4+: StreamingResponse skips its own listen_for_disconnect."""
-    app = FastAPI()
+    app = fenced_app()
 
     @app.get("/work", dependencies=[Depends(disconnect_fencing)])
     async def work() -> StreamingResponse:
@@ -314,7 +352,7 @@ async def test__streaming_response__when_server_owns_no_reader__then_fence_cance
 
 
 async def test__sync_handler__when_fenced__then_ambient_codes_visible() -> None:
-    app = FastAPI()
+    app = fenced_app()
 
     @app.get("/work", dependencies=[Depends(disconnect_fencing)])
     def work() -> dict[str, Any]:
@@ -325,7 +363,7 @@ async def test__sync_handler__when_fenced__then_ambient_codes_visible() -> None:
 
 async def test__sync_handler__when_entering_fence__then_raises_runtime_error() -> None:
     """The threadpool has no running loop, so a `def` handler cannot be fenced."""
-    app = FastAPI()
+    app = fenced_app()
 
     @app.get("/work", dependencies=[Depends(disconnect_fencing)])
     def work() -> dict[str, Any]:
@@ -347,7 +385,7 @@ async def test__sync_handler__when_entering_fence__then_raises_runtime_error() -
 
 async def test__disconnect_fencing__when_code_in_query_string__then_ignored() -> None:
     """A `code` kwarg on the dependency would be a client-settable query param."""
-    app = FastAPI(dependencies=[Depends(disconnect_fencing)])
+    app = fenced_app(dependencies=[Depends(disconnect_fencing)])
 
     @app.get("/work")
     async def work() -> dict[str, Any]:
@@ -359,7 +397,7 @@ async def test__disconnect_fencing__when_code_in_query_string__then_ignored() ->
 
 
 async def test__disconnect_fencing__when_route_declared__then_no_query_param_in_schema() -> None:
-    app = FastAPI()
+    app = fenced_app()
 
     @app.get("/work", dependencies=[Depends(disconnect_fencing)])
     async def work() -> dict[str, Any]:
